more comprehensive makefile (from root directory) support.  implement systemd service.

+++ scripts/mksymlinks.sh
@@ -6,11 +6,16 @@ scriptdir=$(dirname "$(readlink -f "$0")")
 rootdir="$(dirname "$scriptdir")"
 
 # set web root
-webroot=$rootdir/webroot
-ln -s $webroot /var/www/html/gardener
-ln -s $rootdir/index.php /var/www/html/index.php
+webrootdir=$rootdir/webroot
+if [ ! -L /var/www/html/gardener ]; then
+	ln -s $webrootdir /var/www/html/gardener
+fi
 
-echo $rootdir
-echo $scriptdir
-echo $webroot
+if [ ! -L /var/www/html/index.php ]; then
+	ln -s $rootdir/index.php /var/www/html/index.php
+fi
+
+#echo $rootdir
+#echo $scriptdir
+#echo $webrootdir
